feat(log): Support scoped loggers
This patch adds a utility function to optionally embed a scope (and a
scope filter) in the logger. We use a `zerolog.Hook` to dynamically add
a scope field to log events (and filter by scopes).

A common application for this would be running multiple independent
systems in a single application, with the ability to debug with verbose
log output, while discarding logs of certain systems:

  runSubsytemA(log.WithScopedLogger(ctx, "sys.a"))
  runSubsytemB(log.WithScopedLogger(ctx, "sys.b"))

Future patches could extend this to also include functionality to set
log level per scoped logger.

Signed-off-by: Robert Günzler <robert@unikraft.cloud>