feat(protoc-gen-go-gin): Save request body and restore service
This commit makes it possible for an implementing service to read
the request body.  Without this change, the subsequent
`g.ShouldBind` method will read the request body and serialize it
into the input message type; leaving the `g.Request.Body` ready
empty.  Since `Body` does not implement `io.Seeker`, it is not
possible to reset the offset.  Instead, we save the bytes into a
temporary buffer and simply restore it to the `Body` before
invoking the implementing method.  This makes it possible for
the implementing interface method to read the body also.

Signed-off-by: Alexander Jung <alex@unikraft.com>